Trello: ContextList access refactoring, fix but for task menu on edit
