+++ src/types/CommonTypes.ts
@@ -27,7 +27,6 @@ export type TTaskContextType = {
 export type TTaskProps = {
   key: React.Key;
   task: TTask;
-  updateTask: (updatedItem: TTask) => void;
   isFocusOnNew: boolean;
   changeTaskPosition: (listId: number, offset: number) => void;
 };
@@ -36,14 +35,11 @@ export type TListProps = {
   key: React.Key;
   list: TList;
   isFocusOnNewList: boolean;
-  updateList: (updatedItem: TList) => void;
-  changePosition: (listId: number, offset: number) => void;
 };
 
 export type TListHeaderProps = {
   list: TList;
   isFocusOnNewList: boolean;
-  updateList: (updatedItem: TList) => void;
 };
 
 export type TListDetailsProps = {
@@ -53,8 +49,7 @@ export type TListDetailsProps = {
 };
 
 export type TListMoveNavigationProps = {
-  listId: number, 
-  changePosition: (listId: number, offset: number) => void;
+  listId: number;
 };
 
 export type TTaskDetailsProps = {
